requirements.txt
# Core Airflow with GCP and PostgreSQL support
apache-airflow[gcp,postgres]==2.7.3  

# Google Cloud Libraries
google-cloud-storage==2.14.0
google-cloud-bigquery==3.22.0
google-auth==2.25.2  
google-auth-oauthlib==1.1.0  
google-auth-httplib2==0.2.0  

# Data Processing Libraries
pandas==2.1.4  
pyarrow==15.0.0  
fastparquet==2024.2.0  

# PostgreSQL Connector
psycopg2-binary==2.9.9  

# Additional Airflow Dependencies  
apache-airflow-providers-google==10.13.0  
apache-airflow-providers-postgres==5.7.0  
